resolve version specific issues
